requirements.txt
django==2.0.7
djangorestframework==3.8.2
django-rest-framework-social-oauth2==1.1.0
python-dotenv==0.8.2
requests==2.19.1
toolz==0.9.0
